Rewrite coords module in pure Rust
