+++ swiftnav/Cargo.toml
@@ -14,6 +14,7 @@ rustversion = "1.0"
 chrono = { version = "0.4", optional = true }
 swiftnav-sys = { version = "^0.10.0", path = "../swiftnav-sys/" }
 strum = { version = "0.27", features = ["derive"] }
+nalgebra = "0.33"
 thiserror = "2.0"
 
 [dev-dependencies]
