+++ swiftnav/src/ephemeris.rs
@@ -343,7 +343,8 @@ impl Ephemeris {
         // First make sure the ephemeris is valid at `t`, and bail early if it isn't
         self.detailed_status(t).to_result()?;
 
-        let mut sat = AzimuthElevation::default();
+        let mut az = 0.0;
+        let mut el = 0.0;
 
         let result = unsafe {
             swiftnav_sys::calc_sat_az_el(
@@ -351,11 +352,13 @@ impl Ephemeris {
                 &t.to_gps_time_t(),
                 pos.as_array_ref(),
                 swiftnav_sys::satellite_orbit_type_t_MEO,
-                &mut sat.az,
-                &mut sat.el,
+                &mut az,
+                &mut el,
                 true,
             )
         };
+
+        let sat = AzimuthElevation::new(az, el);
 
         assert_eq!(result, 0);
         Ok(sat)
